# core/src/main/resources/openapi/openapi-spec.yml
openapi: 3.0.3
info:
  title: Task by Artur Jarosz
  description: "TASK app openAPI specification"
  version: 0.6.0-SNAPSHOT
...
paths:
...
  /projects/{projectId}/stages/{stageId}/installments:
    post:
      tags:
        - installment
      summary: Create a new installment
      operationId: createInstallment
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: stageId
          in: path
          description: "A unique identifier of the stage"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: New installment data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Installment"
      responses:
        201:
          description: Installment successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
  /projects/{projectId}/installments/{installmentId}:
    get:
      tags:
        - installment
      summary: Get an installment
      operationId: getInstallment
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: installmentId
          in: path
          description: "A unique identifier of the installment"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    put:
      tags:
        - installment
      summary: Update an installment
      operationId: updateInstallment
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: installmentId
          in: path
          description: "A unique identifier of the installment"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated cost data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Installment"
      responses:
        200:
          description: Installment successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    delete:
      tags:
        - installment
      summary: Delete an installment with given installmentId
      operationId: deleteInstallment
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: installmentId
          in: path
          description: "A unique identifier of the installment"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Cost successfully deleted
          content: { }
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
  /projects/{projectId}/installments/{installmentId}/pay:
    post:
      tags:
        - installment
      summary: Pay for an installment
      operationId: payInstallment
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: installmentId
          in: path
          description: "A unique identifier of the installment"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Installement payment data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Installment"
      responses:
        200:
          description: Installment successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
  /projects/{projectId}/installments:
    get:
      tags:
        - installment
      summary: Get installments for project with given projectId
      operationId: getInstallmentsForProject
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installments"
  /projects/{projectId}/supplies:
    post:
      tags:
        - supply
      summary: Create a new supply
      operationId: createSupply
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: New supply data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Supply"
      responses:
        201:
          description: Supply successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    get:
      tags:
        - supply
      summary: Get supplies for project with given projectId
      operationId: getSuppliesForProject
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplies"
  /projects/{projectId}/supplies/{supplyId}:
    delete:
      tags:
        - supply
      summary: Delete a supply with given supplyId
      operationId: deleteSupply
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: supplyId
          in: path
          description: "A unique identifier of the contractor job"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Supply successfully deleted
          content: { }
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    put:
      tags:
        - supply
      summary: Update a supply
      operationId: updateSupply
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: supplyId
          in: path
          description: "A unique identifier of the supply"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated supply data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Supply"
      responses:
        200:
          description: Supply successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    get:
      tags:
        - supply
      summary: Get a supply
      operationId: getSupply
      parameters:
        - name: projectId
          in: path
          description: "A unique identifier of the project"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: supplyId
          in: path
          description: "A unique identifier of the supply"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
...
components:
  schemas:
...
    Installment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        value:
          type: number
        paymentDate:
          type: string
          format: date
        stageId:
          type: integer
          format: int64
        hasInvoice:
          type: boolean
        paid:
          type: boolean
        note:
          type: string
    Installments:
      type: array
      items:
        $ref: "#/components/schemas/Installment"
    Supply:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        value:
          type: number
        supplierId:
          type: integer
          format: int64
        note:
          type: string
        hasInvoice:
          type: boolean
        payable:
          type: boolean
        paid:
          type: boolean
        projectId:
          type: integer
          format: int64
    Supplies:
      type: array
      items:
        $ref: "#/components/schemas/Supply"
...
    ValidatorError:
      type: object
      properties:
        message:
          type: string
...